requirements.txt
accelerate>=0.34.0
bitsandbytes>=0.43.0
datasets>=2.20.0
matplotlib>=3.8.0
numpy>=1.26.0
peft>=0.18.0
pyyaml>=6.0.1
scipy>=1.12.0
torch>=2.2.0
transformers>=4.45.0
trl>=0.10.0
